Minor clarification in README on routes re: propagation
The original explanation's 'even though' is misleading, giving you the impression that if the first `.then` included an error handler, the second `.then`'s error handler would have fired, which only the case if the first error handler threw an error or returned RSVP.reject.

Perhaps we'll want another paragraph explaining how to "rethrow" a reject, but for now this should do.