Properly detect Node environment.
This has been discussed in
[rsvp](https://github.com/tildeio/rsvp.js/pull/401).

In addition to verifying that `process` exists, we need to ensure
that `self` is undefined before declaring that the code is executing in
Node. This will allow `asap` to select the correct `scheduleFlush` method
when running in environments that also define `process` in the browser (such
as NW.js or Electron).